refactor: modularize shop editor
